+++ CHANGELOG.md
@@ -6,8 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+
+## [1.0.2] - 2022-10-27
+
 ### Added
+- **Tests** - Test for `ssh.py` and `rsync.py` module.
 - **GitHub** - Workflow to run unittests on every PR and push to main.
+
+### Changed
+- **SSH** - Add error to show when SSH credentials are invalid.
+- **Dependencies** - Add `coverage` as a test dependency.
+- **Readme** - Add new badges to the Readme.
+- **Common** - Move EC2 pricing calls to single function in `common.py`.
 
 
 ## [1.0.1] - 2022-09-28
@@ -23,12 +33,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - **GitHub** - Update action to build and publish package only when version is bumped.
 - **Forge** - Added automatic tag `forge-name` to allow `Name` tag to be changed.
 
-
 ## [1.0.0] - 2022-09-27
 
 ### Added
 - **Initial commit** - Forge source code, unittests, docs, pyproject.toml, README.md, and LICENSE files.
 
-[unreleased]: https://github.com/carsdotcom/cars-forge/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/carsdotcom/cars-forge/compare/v1.0.2...HEAD
+[1.0.2]: https://github.com/carsdotcom/cars-forge/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/carsdotcom/cars-forge/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/carsdotcom/cars-forge/releases/tag/v1.0.0